requirements.txt
google-adk==1.21.0
fastapi
litellm
fastmcp==2.10.6
neo4j==5.28.1
sentence_transformers==4.1.0
tiktoken==0.9.0
tqdm==4.67.1
tree_sitter==0.25.2
tree_sitter_java==0.23.5
tree_sitter_kotlin==1.1.0

